quick commit

+++ localsuppress2014/sfr_m200.py
@@ -95,7 +95,7 @@ def plot_uv_z8():
         gal[index] = gal[index][numpy.where(gal[index]["HaloM_Crit200"]>0.)]
         nummax= numpy.nanmax(gal[index]["NPhotReion"])
         gal[index]["NPhotReion"] = numpy.clip(gal[index]["NPhotReion"]+numpy.log10(SEC_PER_YEAR),0.0,nummax)
-        print numpy.log10(gal[index]["HaloM_Crit200"]*1.e10*Msun2kg/h_mass)
+        print numpy.log10(gal[index]["HaloM_Crit200"].astype(numpy.float64)*1.e10*Msun2kg/h_mass)
         sum_logphoton[index] = numpy.histogram(numpy.log10(gal[index]["HaloM_Crit200"]*1.e10),range=rangen,bins=bins,weights=gal[index]["NPhotReion"]-numpy.log10(gal[index]["HaloM_Crit200"]*1.e10*Msun2kg/h_mass*0.167))
         ssfr = gal[index]["Sfr"]/(gal[index]["HaloM_Crit200"]*1.e10)
         ssfr = numpy.nan_to_num(ssfr)
